backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
pillow==10.4.0